move screenshot test, update package in fastlane
